new infrastructure. favors separate development and circular dependency avoidance
