+++ forsyde-atom.cabal
@@ -34,23 +34,31 @@ license:             BSD3
 license-file:        LICENSE
 author:              George Ungureanu
 maintainer:          ugeorge@kth.se
-copyright:           Copyright (c) 2015, George Ungureanu, KTH/ICT/ESY
+copyright:           Copyright (c) 2015-2016, George Ungureanu, KTH/ICT/ESY
 homepage:            https://github.com/forsyde/forsyde-atom
-category:            model, cyber-physical systems, skeletons
+category:            model, cyber-physical systems, shallow-embedded
 build-type:          Simple
 cabal-version:       >=1.22
 extra-doc-files:     includes/figs/*.png
 
 library
-  exposed-modules:     ForSyDe
-                       ForSyDe.Core,
-                       ForSyDe.Core.Signal,
-                       ForSyDe.Core.Vector,
-                       ForSyDe.Core.MoC,
-                       ForSyDe.Core.ValueExt,
-                       ForSyDe.Core.Utilities,
-                       ForSyDe.MoCLib.SY,
-                       ForSyDe.MoCLib.DT
+  exposed-modules:     ForSyDe.Atom
+                       ForSyDe.Atom.Behavior,
+                       ForSyDe.Atom.Behavior.Atoms,
+                       ForSyDe.Atom.Behavior.Constructors,
+                       ForSyDe.Atom.Behavior.ValueExt,
+                       ForSyDe.Atom.MoC,
+                       ForSyDe.Atom.MoC.Signal,
+                       ForSyDe.Atom.MoC.Atoms,
+                       ForSyDe.Atom.MoC.Constructors,
+                       ForSyDe.Atom.MoC.Interfaces,
+                       ForSyDe.Atom.MoC.SY,
+                       ForSyDe.Atom.MoC.SY.Core,
+                       ForSyDe.Atom.MoC.SY.Lib,
+                       ForSyDe.Atom.MoC.DE,
+                       ForSyDe.Atom.MoC.DE.Core,
+                       ForSyDe.Atom.MoC.DE.Lib,
+                       ForSyDe.Atom.Utilities
   -- other-modules:       
   Default-Language:    Haskell2010
   other-extensions:    PostfixOperators
