+++ src/ForSyDe/Atom/MoC/CT/Core.hs
