+++ src/ForSyDe/Atom/MoC/Constructors.hs
@@ -1,8 +1,8 @@
 {-# LANGUAGE PostfixOperators #-}
-{-# OPTIONS_HADDOCK show-extensions #-}
+{-# OPTIONS_HADDOCK show-extensions, prune, hide #-}
 -----------------------------------------------------------------------------
 -- |
--- Module      :  ForSyDe.Core.Signal
+-- Module      :  ForSyDe.Atom.MoC.Constructors
 -- Copyright   :  (c) George Ungureanu, KTH/ICT/ESY 2015
 -- License     :  BSD-style (see the file LICENSE)
 -- 
@@ -14,180 +14,12 @@
 -- synchronization layer.
 -----------------------------------------------------------------------------
 
-module ForSyDe.Core.MoC(
-  -- * Atoms
-  
-  MoC(..),
+module ForSyDe.Atom.MoC.Constructors where
 
-  -- * Process constructors
+import ForSyDe.Atom.MoC.Atoms
+import ForSyDe.Atom.Utilities
+import ForSyDe.Atom.Behavior
 
-  -- | As shown in the documentation of "ForSyDe.Core" process
-  -- constructors are implemented as compositions of MoC atoms. In
-  -- order to avoid working with signals of tuples @unzip@ and in
-  -- order for the process network to reflect the passed functions,
-  -- process utilities (defined in "ForSyDe.Core.Utilities") are also
-  -- used.
-  --
-  -- Due to Haskell's strict type system and the implementation
-  -- mechanisms, we need to provide separate constructors @nameXY@,
-  -- where @name@ is the process constructor type, @X@ is the number
-  -- of inputs and @Y@ is the number of outputs. This module provides
-  -- constructors with /X &#8804; 4/ and /Y &#8804; 4/. If needed, the
-  -- designer is free to implement her own constructor by following
-  -- the atom composition rules in the source code.
-
-  -- ** @comb@
-
-  -- | @comb@ processes map combinatorial functions on signals and
-  -- take care of synchronization between input signals.
-  --
-  -- <<includes/figs/comb-formula.png>>
-  -- <<includes/figs/comb-graph.png>>
-
-  comb11, comb12, comb13, comb14,
-  comb21, comb22, comb23, comb24,
-  comb31, comb32, comb33, comb34,
-  comb41, comb42, comb43, comb44,
-  comb51, comb52, comb53, comb54,
-
-  -- ** @delay@
-
-  -- | The @delay@ process provides both an initial token and shifts the
-  -- phase of the signal. In other words, it "delays" a signal with
-  -- one event.
-  --
-  -- <<includes/figs/delay-formula.png>>
-  -- <<includes/figs/delay-graph.png>>
-  delay, (-&>-),
-
-  -- ** @scanl@
-
-  -- | @scanl@ processes model generates the graph shown below. There
-  -- exists a variant with 0 input signals, in which case the process
-  -- is a signal generator.
-  --
-  -- <<includes/figs/scanl-formula.png>>
-  -- <<includes/figs/scanl-graph.png>>
-
-  -- scanl01, scanl02, scanl03, scanl04,
-  scanl11, scanl12, scanl13, scanl14,
-  scanl21, scanl22, scanl23, scanl24,
-  scanl31, scanl32, scanl33, scanl34,
-  scanl41, scanl42, scanl43, scanl44,
-
-  
-  -- ** @scanld@
-
-  -- | @scanld@ processes model generates the graph shown below. There
-  -- exists a variant with 0 input signals, in which case the process
-  -- is a signal generator.
-  --
-  -- <<includes/figs/scanld-formula.png>>
-  -- <<includes/figs/scanld-graph.png>>
-
-  scanld01, scanld02, scanld03, scanld04,
-  scanld11, scanld12, scanld13, scanld14,
-  scanld21, scanld22, scanld23, scanld24,
-  scanld31, scanld32, scanld33, scanld34,
-  scanld41, scanld42, scanld43, scanld44,
-  
-  -- ** @moore@
-
-  -- | @moore@ processes model Moore state machines.
-  --
-  -- <<includes/figs/moore-formula.png>>
-  -- <<includes/figs/moore-graph.png>>
-
-  moore11, moore12, moore13, moore14,
-  moore21, moore22, moore23, moore24,
-  moore31, moore32, moore33, moore34,
-  moore41, moore42, moore43, moore44,
-
-  -- ** @mealy@
-
-  -- | @mealy@ processes model Mealy state machines.
-  --
-  -- <<includes/figs/mealy-formula.png>>
-  -- <<includes/figs/mealy-graph.png>>
-
-  mealy11, mealy12, mealy13, mealy14,
-  mealy21, mealy22, mealy23, mealy24,
-  mealy31, mealy32, mealy33, mealy34,
-  mealy41, mealy42, mealy43, mealy44,  
-  
-  ) where
-
-import ForSyDe.Core.Signal
-import ForSyDe.Core.Utilities
-import ForSyDe.Core.ValueExt
-
-infixl 5 -$-, -*-
-infixl 3 ->-, -&-
-
--- | This is a type class holding the synchronization layer atoms.
--- Each model of computation defines its tag system by providing an
--- @Event@ type constructor (that defines /T/ &#215; /V/) which is an
--- instance of this class.
---
--- In order to simplify the mathematical notations, we shall consider
---the following function as the type constructor for a MoC-bound
---signal:
---
--- <<includes/figs/signal-constructor-formula.png>>
---
-class MoC e where
-
-  -- | This is a rough equivalent of the @(\<$\>)@ functor operator,
-  -- mapping a function on extended values to a MoC-bound signal
-  -- (signal of tagged events). Its mathematical definition is:
-  --
-  -- <<includes/figs/star-atom-formula.png>>
-  --
-  -- The reason why &#946; is not extended is to allow for the
-  -- composition of generic process constructors with arbitrary number
-  -- of arguments.
-  (-$-)  :: (Value a -> b) -> Signal (e (Value a)) -> Signal (e b)
-
-  -- | This is a rough equivalent of the @(\<*\>)@ applicative functor
-  -- operator, used for synchronizing and applying a signal of
-  -- functions (bound to a MoC) with a signal of values (bound to the
-  -- same MoC). Its mathematical signature is:
-  --
-  -- <<includes/figs/ostar-atom-formula.png>>
-  --
-  -- The reason why &#946; is not extended is to allow for the
-  -- composition of generic process constructors with arbitrary number
-  -- of arguments.
-  (-*-) :: Signal (e (Value a -> b)) -> Signal (e (Value a)) -> Signal (e b)
-
-  -- | Since ForSyDe signals are modeled similar to
-  -- <https://www.cs.ox.ac.uk/files/3378/PRG56.pdf Bird lists>, the
-  -- inherent execution model of ForSyDe process networks is the
-  -- dataflow network (Kahn process network), as shown by Reekie in
-  -- his <http://ptolemy.eecs.berkeley.edu/~johnr/papers/thesis.html process nets>.
-  -- Therefore it is of utmost importance to provide an atom which
-  -- provides an initial event (token) to avoid execution deadlock in
-  -- feedback loops (self-referential processes). Concretely, the
-  -- @-\>-@ atom prepends an event at the head of a signal. Its
-  -- signature used in mathematical formulas is:
-  --
-  -- <<includes/figs/pre-atom-formula.png>>
-  (->-) :: e a -> Signal (e a) -> Signal (e a)
-   
-  -- | Another property, this time derived from the tagged signal
-  -- model, requires the processes to be /monotonic/
-  -- (order-preserving) in order to preserve determinancy. Therefore
-  -- we introduce the @(-&-)@ atom as means to manipulate the tags in
-  -- a signal in a manner which respects monotonicity. Its behavior
-  -- could be described as "shifting the phase of a signal with a
-  -- positive constant", thus preserving its characteristic function
-  -- intact. Its signature used in mathematical formulas is:
-  --
-  -- <<includes/figs/phi-atom-formula.png>>
-  (-&-) :: e a -> Signal (e a) -> Signal (e a)
-
-           
------------------------------------------------------------------------------
 comb11 f s1             = (f -$- s1)
 comb12 f s1             = (f -$- s1 ||<)
 comb13 f s1             = (f -$- s1 ||<<)
