+++ src/ForSyDe/Atom/Skeleton.hs
