+++ src/ForSyDe/Atom/Skeleton/Patterns.hs
