+++ src/ForSyDe/Atom/Skeleton/Transition.hs
