+++ src/ForSyDe/Atom/Utilities.hs
@@ -15,7 +15,7 @@
 -- does not suffice.
 -----------------------------------------------------------------------------
 
-module ForSyDe.Core.Utilities(
+module ForSyDe.Atom.Utilities(
   -- * @at@ functions
   
   -- | The @at@/xy/ functions return the /y/-th element of an
@@ -193,3 +193,14 @@ infixl 3 |||<,  |||<<, |||<<<, |||<<<<, |||<<<<<, |||<<<<<<, |||<<<<<<<, |||<<<<
 (|||<<<<<<)   s = (((|<<<<<<) <$>) <$> s   ||<<<<<<)
 (|||<<<<<<<)  s = (((|<<<<<<<) <$>) <$> s  ||<<<<<<<)
 (|||<<<<<<<<) s = (((|<<<<<<<<) <$>) <$> s ||<<<<<<<<)
+
+
+infixl 6 <>,  <>>, <>>>, <>>>>, <>>>>>, <>>>>>>, <>>>>>>>, <>>>>>>>>
+(<>)        (a1, a2)                             = a1 a2
+(<>>)       (a1, a2, a3)                         = a1 a2 a3
+(<>>>)      (a1, a2, a3, a4)                     = a1 a2 a3 a4
+(<>>>>)     (a1, a2, a3, a4, a5)                 = a1 a2 a3 a4 a5
+(<>>>>>)    (a1, a2, a3, a4, a5, a6)             = a1 a2 a3 a4 a5 a6
+(<>>>>>>)   (a1, a2, a3, a4, a5, a6, a7)         = a1 a2 a3 a4 a5 a6 a7
+(<>>>>>>>)  (a1, a2, a3, a4, a5, a6, a7, a8)     = a1 a2 a3 a4 a5 a6 a7 a8
+(<>>>>>>>>) (a1, a2, a3, a4, a5, a6, a7, a8, a9) = a1 a2 a3 a4 a5 a6 a7 a8 a9
